Move post_tool_use + precommit_runner + test under hook_daemon/
Part of consolidating hook implementations into the hook_daemon package
where they are actually dispatched.
