+++ devinfra/claude/BUILD.bazel
@@ -254,28 +254,6 @@ py_library(
 )
 
 py_library(
-    name = "precommit_runner",
-    srcs = ["precommit_runner.py"],
-    imports = ["../.."],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//util:fs",
-        "@pypi//pre_commit",
-    ],
-)
-
-py_library(
-    name = "post_tool_use_lib",
-    srcs = ["post_tool_use.py"],
-    imports = ["../.."],
-    visibility = ["//visibility:public"],
-    deps = [
-        ":precommit_runner",
-        "//devinfra/claude/claude_api/hooks:post_tool_use",
-    ],
-)
-
-py_library(
     name = "hook_dispatch_lib",
     srcs = ["hook_dispatch.py"],
     imports = ["../.."],
@@ -377,20 +355,6 @@ py_library(
         ":session_paths",
         ":settings",
         "@pypi//pytest",
-    ],
-)
-
-py_test(
-    name = "test_post_tool_use",
-    srcs = ["test_post_tool_use.py"],
-    imports = ["../.."],
-    deps = [
-        ":conftest",
-        ":post_tool_use_lib",
-        ":precommit_runner",
-        "//devinfra/claude/claude_api/hooks:post_tool_use",
-        "@pypi//pytest",
-        "@pypi//pytest_bazel",
     ],
 )
