+++ devinfra/claude/hook_daemon/BUILD.bazel
@@ -23,14 +23,36 @@ py_library(
 )
 
 py_library(
+    name = "precommit_runner",
+    srcs = ["precommit_runner.py"],
+    imports = ["../../.."],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//util:fs",
+        "@pypi//pre_commit",
+    ],
+)
+
+py_library(
+    name = "post_tool_use_lib",
+    srcs = ["post_tool_use.py"],
+    imports = ["../../.."],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":precommit_runner",
+        "//devinfra/claude/claude_api/hooks:post_tool_use",
+    ],
+)
+
+py_library(
     name = "server",
     srcs = ["server.py"],
     imports = ["../../.."],
     visibility = ["//visibility:public"],
     deps = [
         ":models",
+        ":post_tool_use_lib",
         ":pre_tool_use_lib",
-        "//devinfra/claude:post_tool_use_lib",
         "//devinfra/claude:session_paths",
         "//devinfra/claude:session_start_lib",
         "//devinfra/claude:settings",
@@ -118,6 +140,20 @@ py_test(
         ":pre_tool_use_lib",
         "//devinfra/claude:conftest",
         "//devinfra/claude/claude_api/hooks:pre_tool_use",
+        "@pypi//pytest",
+        "@pypi//pytest_bazel",
+    ],
+)
+
+py_test(
+    name = "test_post_tool_use",
+    srcs = ["test_post_tool_use.py"],
+    imports = ["../../.."],
+    deps = [
+        ":post_tool_use_lib",
+        ":precommit_runner",
+        "//devinfra/claude:conftest",
+        "//devinfra/claude/claude_api/hooks:post_tool_use",
         "@pypi//pytest",
         "@pypi//pytest_bazel",
     ],
